services/fastapi-recommendation-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
kafka-python==2.0.2
redis==5.0.1
pymongo==4.6.0
motor==3.3.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
prometheus-client==0.19.0
httpx==0.25.2
black==23.12.0
flake8==6.1.0
isort==5.13.2
python-dotenv
schedule
requests
numpy
scikit-learn
